No_Photo_Zone/requirements.txt
mysql-connector-python
requests
opencv-python
urllib3
schedule